Fix cjmp sources handling and diagnostics

Created-by: mashser
Commit-by: Mashkov Sergey
Merged-by: CangjiePL
Description: ## Change Details (Required)

Fix cjmp sources handling and diagnostics

Fix assertions in the compiler related to fileID and diagnostics related to cjo package mismatch 

## Change Type (Required)

- [ ] Feature
- [x] Bugfix
- [ ] Build Process or Auxiliary Tool Changes
- [ ] Documentation Update

## Self-Check of Changes (Required)

### Local Compilation Verification Results:

- [x] Cangjie Compiler compiled successfully
- [x] Cangjie Runtime compiled successfully
- [x] Cangjie Standard Library compiled successfully
- [ ] Not applicable (select this option for auxiliary tool changes or documentation updates)

### Local Test Case Verification Results:

- [x] Pass. Please provide screenshots below.
- [ ] Not applicable. Please explain the reasons below.

![изображение.png](https://raw.gitcode.com/user-images/assets/7239667/a289d865-bc43-4f0d-bb2e-507573afe87e/изображение.png 'изображение.png')

## Related Issues (Required)

[Fix cjmp sources handling and diagnostics](https://gitcode.com/Cangjie/cangjie_compiler/issues/635)

See merge request: Cangjie/cangjie_compiler!1053

+++ src/CHIR/Interpreter/CHIR2BCHIR.cpp
@@ -563,7 +563,7 @@ const std::unordered_map<std::string, IntrinsicKind> CHIR2BCHIR::syscall2Intrins
 
 Bchir::CodePosition CHIR2BCHIR::CHIRPos2BCHIRPos(const DebugLocation& loc)
 {
-    auto fileName = sourceManager.GetSource(static_cast<unsigned int>(loc.GetFileID())).path;
+    auto fileName = sourceManager.GetSource(loc.GetFileID()).path;
     auto fileId = fileNameToIndexMemoization.find(fileName);
     size_t fileIdx = 0;
     if (fileId == fileNameToIndexMemoization.end()) {
